Update alert_crypto.py
found out when keeping the crypto specific variables as global variables, it would send texts throughout the day with the same value as the first text, never changing. by putting the variables inside the txt_code function, it becomes a dynamic variable with the most up to date information. seems that the global variables are a one time request to the exchange for value, but when it inside the function it will get the value as often as you define, mine is hourly between 6-midnight.